requirements.txt
akshare>=1.12.0
pandas>=2.0.0
requests>=2.28.0
pyyaml>=6.0
numpy>=1.24.0
streamlit>=1.30.0
